Update README.md
The Backlink Checker Tool is designed to make the process of validating and auditing Backlinks on a webpage faster and more efficient. This tool can be especially useful for SEO professionals, website developers, or anyone managing a website who wants to ensure that internal and external links are working correctly and have the appropriate attributes.